Initial effort
I expanded the example pygls server from the example page to use
thefuzz fuzzy complete to match against the current word and
present the docstring with completion. I haven't yet implemented
the scriptlets (since they have a different format in the manual
page that I need to write a different parser for) but this is off
to a good start.

I'm currently testing it in Kate since Kate makes it pretty easy
to add a custom LSP, you just add a user config like:

{
    "servers": {
        "spec": {
            "command": ["/home/jesingh/code/rpm-spec-lsp-server/src/main.py"],
            "url": "https://git.sr.ht/~mcepl/rpm-spec-lsp-server",
            "highlightingModeRegex": "^RPM Spec$"
        }
    }
}

Still haven't touched on implementing macro checking with
shellcheck or whatever we end up using there. Looking forward to
doing more of this next week.

Signed-off-by: Jehan Singh <jehan.singh@suse.com>